Camel cases YAML file names (#11)
TL;DR
-----

Makes the case of names of the YAML files match the names the
applications use for themselves.

Details
-------

Configurations still aren't loading after simplifying the
repository and gettting the branch right. Playing with the API
directly suggested that the application name that the config
server receives is case sensitive.

Updates the names of all application to match the camel-cased
names that they use for themselves. Every likes being known the
way they prefer.